Edit SimilarImageFinder : Using Gpu
