+++ .gitmodules
@@ -1,4 +1,3 @@
 [submodule "Cross-cutting-with-face/yoloface"]
 	path = Cross-cutting-with-face/yoloface
 	url = https://github.com/elyha7/yoloface.git
-#
