+++ Cross-cutting-with-face/SimilarImageFinder.py
@@ -1,12 +1,10 @@
 from yoloface.face_detector import YoloDetector
-from PIL import Image
-import numpy as np
-import pandas as pd
 from util.distance import findEuclideanDistance
 from itertools import combinations
 from util.VideoCapture import VideoCapture
-from util.functions import get_all_images, plot_by_cv2
-
+from util.plot import plot_by_cv2
+import torch
+from similarImageFinder_functions import *
 
 class SimilarImageFinder:
     def __init__(self, detector, video_name, video_path, period=30, video_capture=True):
@@ -22,86 +20,31 @@ class SimilarImageFinder:
         # 이미지 출력
         plot_by_cv2(similarity_list, video_name)
 
-    def _sort_by_x1(self, boxes):
-        """
-        감지한 boundary box list 를 x1이 작은 순서대로 (왼쪽부터) 정렬하는 함수입니다.
-        """
-        boxes = sorted(boxes, key=lambda x: x[0][0])
-        return boxes
-
-    def _zip_person(self, boxes, points):
-        """
-        box 와 landmark 의 리스트를 각 사람의 box 와 landmark 로 합치는 함수입니다.
-        """
-        person = zip(boxes[0], points[0])
-        return list(person)
-
     def _make_df_detection_by_detector(self, detector, video_name):
         """
         Face Detection 데이터 프레임을 만드는 함수입니다.
         :param: detector ['opencv', 'ssd', 'mtcnn', 'retinaface', 'yoloface]
         :return:
         """
-        df_detection = pd.DataFrame(columns=['frame_num', 'video_num', 'detect_person_num', 'boxes', 'landmarks'])
-        images = get_all_images(video_name) # 모든 이미지 경로를 가져옵니다.
         if detector == 'yoloface':
             try:
+                torch.tensor(1).cuda()
                 yolo_detector = YoloDetector(target_size=720, gpu=1, min_face=90)
                 print('gpu 사용')
+                df_detection = detect_images_by_gpu(yolo_detector, video_name, batch_size=64)
             except:
                 yolo_detector = YoloDetector(target_size=720, gpu=-1, min_face=90)
                 print('cpu 사용')
-            for image in images:
-                # 영상 번호
-                video_num = image.split('/')[-1][:-4]  # [:-4] -> .jpg remove
-                # frame 번호
-                frame_num = image.split('/')[-2]
-                # 이미지 경로 -> ndarray
-                image = np.array(Image.open(image))
-                # get boundary box and landmarks
-                boxes, points = yolo_detector.predict(image)
-                # 감지된 사람 중 왼쪽에 있는 사람 순(x1이 작은 순)으로 정렬
-                people = self._sort_by_x1(self._zip_person(boxes, points))
-                # 1명 이상인 경우 / 1명만 검출하고 싶으면 == 1 로 변경
-                if len(people) > 0:
-                    data = {
-                        'frame_num': int(frame_num),
-                        'video_num': int(video_num),
-                        'detect_person_num': len(people),
-                        'boxes': [person[0] for person in people],
-                        'landmarks': [person[1] for person in people]
-                    }
-                    df_detection = df_detection.append(data, ignore_index=True)
-                    print(f'{frame_num}번 째 frame : {video_num} 영상 이미지 저장')
+                df_detection = detect_images_by_cpu(yolo_detector, video_name)
         else:
             print('아직 미구현, Deepface Detector 는 얼굴을 추출하기 때문에 customizing 필요합니다')
         # frame_num 순으로 정렬
         df_detection = df_detection.sort_values(by='frame_num')
         # filter 된 dataframe
-        df_filtered = self._filter_df(df_detection)
+        df_filtered = filter_df(df_detection)
         # 데이터프레임 저장
         df_filtered.to_parquet(f'./dataset/{video_name}/detection_data/yoloface_data.parquet')
 
-        return df_filtered
-
-    def _filter_df(self, df):
-        """
-        얼굴을 감지한 데이터프레임에서 감지된 사람 수가 같은 데이터만 추출하여 데이터프레임을 만듭니다.
-        """
-        # 같은 프레임의 이미지들 중에서 적어도 2번 이상 사람이 감지된 프레임만 추출
-        frame_index = df['frame_num'].value_counts()[df['frame_num'].value_counts() > 1].index
-        # 'frame_num' index 설정
-        df = df.set_index('frame_num')
-
-        # 빈 df_filtered 생성
-        df_filtered = pd.DataFrame(
-            columns=['frame_num', 'video_num', 'detect_person_num', 'boxes', 'landmarks']
-        ).set_index('frame_num')
-
-        for frame_num in frame_index:
-            df1 = df.loc[frame_num]
-            mask = df1.loc[frame_num].duplicated(['detect_person_num'], keep=False)
-            df_filtered = pd.concat([df_filtered, df1.loc[mask]])
         return df_filtered
 
     def find_similarity_images(self, df_filtered):
@@ -132,7 +75,7 @@ class SimilarImageFinder:
 
 
 if __name__ == '__main__':
-    video_path = '/Users/jungsuplim/Desktop/origin_video'
+    video_path = 'C:/Users/JungSupLim/Desktop/video'
     # video capture 생략하고 싶으면 False
-    SimilarImageFinder('yoloface', 'iu_lilac3', video_path, period=900, video_capture=False)
+    SimilarImageFinder('yoloface', 'idle_tomboy', video_path, period=10, video_capture=True)
