+++ Cross-cutting-with-face/requirements.txt
@@ -3,6 +3,7 @@ scipy==1.7.2
 seaborn==0.11.1
 joblib==1.0.1
 matplotlib==3.4.3
+fastparquet==2022.11.0
 torch==1.10.1+cu113; sys_platform == 'win32' 
 torch==1.10.1; sys_platform == 'darwin'
 pandas==1.3.4
@@ -16,3 +17,4 @@ onnx==1.10.2
 Pillow==9.0.0
 PyYAML==6.0
 wandb==0.12.9
+
